Cleanup theme and application bar
