+++ bookchain-client/src/components/ApplicationBar/index.js
@@ -1,0 +1,3 @@
+import ApplicationBar from './ApplicationBar';
+
+export default ApplicationBar;
