+++ bookchain-client/src/components/Theme/Theme.js
@@ -1,0 +1,10 @@
+import { createMuiTheme } from '@material-ui/core/styles';
+import teal from '@material-ui/core/colors/teal';
+
+const Theme = createMuiTheme({
+  palette: {
+    secondary: teal
+  }
+});
+
+export default Theme;
