+++ bookchain-client/src/components/Theme/index.js
@@ -1,0 +1,3 @@
+import Theme from './Theme';
+
+export default Theme;
